+++ mayan/apps/events/views/mixins.py
@@ -1,7 +1,19 @@
 from django.http import Http404
 
+from mayan.apps.acls.models import AccessControlList
+
 from ..classes import EventType
 from ..literals import TEXT_UNKNOWN_EVENT_ID
+from ..permissions import permission_events_view
+
+
+class NotificationViewMixin:
+    def get_source_queryset(self):
+        return AccessControlList.objects.restrict_queryset(
+            permission=permission_events_view,
+            queryset=self.request.user.notifications.all(),
+            user=self.request.user
+        )
 
 
 class VerbEventViewMixin: