+++ mayan/apps/events/views/notification_views.py
@@ -5,15 +5,18 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.views.generics import ConfirmView, SingleObjectListView
 
-from ..icons import icon_notification_list
+from ..icons import (
+    icon_notification_list, icon_notification_mark_read,
+    icon_notification_mark_read_all
+)
 from ..links import link_event_type_subscription_list
 
-__all__ = (
-    'NotificationListView', 'NotificationMarkRead', 'NotificationMarkReadAll'
-)
+from .mixins import NotificationViewMixin
 
 
-class NotificationListView(SingleObjectListView):
+class NotificationListView(NotificationViewMixin, SingleObjectListView):
+    view_icon = icon_notification_list
+
     def get_extra_context(self):
         return {
             'hide_object': True,
@@ -27,28 +30,23 @@ class NotificationListView(SingleObjectListView):
             ),
             'no_results_title': _('There are no notifications'),
             'object': self.request.user,
-            'title': _('Notifications'),
+            'title': _('Notifications')
         }
 
-    def get_source_queryset(self):
-        return self.request.user.notifications.all()
 
-
-class NotificationMarkRead(ConfirmView):
+class NotificationMarkRead(NotificationViewMixin, ConfirmView):
     post_action_redirect = reverse_lazy(
         viewname='events:user_notifications_list'
     )
+    view_icon = icon_notification_mark_read
 
     def get_extra_context(self):
         return {
             'title': _('Mark the selected notification as read?')
         }
 
-    def get_queryset(self):
-        return self.request.user.notifications.all()
-
     def view_action(self, form=None):
-        self.get_queryset().filter(
+        self.get_source_queryset().filter(
             pk=self.kwargs['notification_id']
         ).update(read=True)
 
@@ -57,21 +55,19 @@ class NotificationMarkRead(ConfirmView):
         )
 
 
-class NotificationMarkReadAll(ConfirmView):
+class NotificationMarkReadAll(NotificationViewMixin, ConfirmView):
     post_action_redirect = reverse_lazy(
         viewname='events:user_notifications_list'
     )
+    view_icon = icon_notification_mark_read_all
 
     def get_extra_context(self):
         return {
             'title': _('Mark all notification as read?')
         }
 
-    def get_queryset(self):
-        return self.request.user.notifications.all()
-
     def view_action(self, form=None):
-        self.get_queryset().update(read=True)
+        self.get_source_queryset().update(read=True)
 
         messages.success(
             message=_('All notifications marked as read.'),